backend/requirements.txt
# Web scraping
requests
beautifulsoup4
selenium
cloudscraper
python-Levenshtein

# Scheduler
schedule

# Email handling
sendgrid

# Firebase database
firebase-admin

# Chrome WebDriver manager
webdriver-manager

# Environment Variables
python-dotenv

# Data handling
geopy